+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [DOCS] Clarify when `format` throws a `RangeError` and fix the test for it
 - [ENHANCEMENT] More extensive validation of numeric time zone offsets
 - [BUGFIX] Fixed `zonedTimeToUtc` throwing `RangeError` instead of returning an Invalid Date (#151)
+- [BUGFIX] Fixed `format` not throwing `RangeError` for invalid time zones with offset tokens (#152)
 
 ### v1.2.2 (21 December 2021)